+++ packages/core/src/languages/css/__tests__/css-values.test.ts
@@ -43,6 +43,27 @@ suite("CSS - CSS Value Expression Factory", function() {
         },
       ],
     },
+    {
+      name: "multi-value",
+      cases: [
+        {
+          input: "div { margin: 0 42px; }",
+          pattern: "[0-9]+",
+          expected: ["42"],
+          options: {
+            suffix: "px",
+          },
+        },
+        {
+          input: "div { padding: 0 3px 0 14px; }",
+          pattern: "[0-9]+",
+          expected: ["3", "14"],
+          options: {
+            suffix: "px",
+          },
+        },
+      ],
+    },
   ];
 
   for (const { name, cases } of scenarios) {